\name{get_Q_1indiv}
\alias{get_Q_1indiv}
\title{Formulate an RW1 Q matrix (precision matrix = tau*Q) for one individual}
\usage{
  get_Q_1indiv(n)
}
\arguments{
  \item{n}{number of timesteps an individual is available
  to be observed}
}
\value{
  returns an (n x n) Q matrix
}
\description{
  Formulate an RW1 Q matrix (precision matrix = tau*Q) for
  one individual
}
\author{
  Paul Conn
}
\keyword{CAR,}
\keyword{matrix}
\keyword{Q}
\keyword{RW1,}

