\name{get_Q_1indiv_RW2}
\alias{get_Q_1indiv_RW2}
\title{Formulate an RW2 Q matrix (precision matrix = tau*Q) for one individual}
\usage{
  get_Q_1indiv_RW2(n)
}
\arguments{
  \item{n}{number of timesteps an individual is available
  to be observed}
}
\description{
  Formulate an RW2 Q matrix (precision matrix = tau*Q) for
  one individual
}
\author{
  Paul Conn
}
\keyword{CAR,}
\keyword{matrix}
\keyword{Q}
\keyword{RW2,}

