\name{rrw}
\alias{rrw}
\title{Simulate an ICAR Process}
\usage{
  rrw(x)
}
\arguments{
  \item{x}{a precision matrix = tau*Q}
}
\value{
  returns a vector of spatially correlated random effects
  of length = to the number of rows of x
}
\description{
  Simulate an ICAR Process
}
\author{
  Devin Johnson
}
\keyword{CAR,}
\keyword{simulation}

